+++ aemedge/blocks/cards/cards.css
@@ -1,27 +1,63 @@
-.cards > ul {
-  list-style: none;
-  margin: 0;
-  padding: 0;
+.cards-wrapper .cards .cards-group-container .grouped-content-container {
   display: grid;
-  grid-template-columns: repeat(auto-fill, minmax(257px, 1fr));
-  grid-gap: 24px;
+  grid-template-columns: 1fr;
+  grid-auto-rows: auto;
+  gap: 40px;
 }
 
-.cards > ul > li {
-  border: 1px solid #dadada;
-  background-color: var(--background-color);
+.cards-group-container .grouped-content-container .grouped-content {
+  border-left: 7px solid;
+  padding-left: 15px;
+  border-left-color: unset;
+  box-sizing: border-box;
 }
 
-.cards .cards-card-body {
-  margin: 16px;
+.cards-group-container .grouped-content-container .grouped-content p {
+  margin-top: 0;
+  font-family: Roboto, 'Roboto Fallback', sans-serif;
+  font-size: 14.4px;
+  color: #696969;
 }
 
-.cards .cards-card-image {
-  line-height: 0;
+.cards-group-container .grouped-content-container .grouped-content p:first-child {
+  font-weight: bold;
+  font-size: 22.4px;
+  color: var(--alt-link-hover-color);
 }
 
-.cards > ul > li img {
-  width: 100%;
-  aspect-ratio: 4 / 3;
-  object-fit: cover;
+.cards-group-container .grouped-content-container .grouped-content p:first-child a {
+  color: var(--alt-link-color);
+  text-decoration: underline;
+  font-weight: bold;
+  font-size: 22.4px;
 }
+
+.cards-group-container .grouped-content-container .grouped-content p a:hover {
+  text-decoration: none;
+  color: var(--alt-link-hover-color);
+}
+
+.grouped-content-container .grouped-content.blue {
+  border-left-color: #00607F;
+}
+
+.grouped-content-container .grouped-content.red {
+  border-left-color: #BB1F53;
+}
+
+.grouped-content-container .grouped-content.grey {
+    border-left-color: #B9C8D3;
+}
+
+@media (width >= 900px) {
+  .cards-wrapper .cards {
+    max-width: 950px;
+    margin: auto auto 40px 17%;
+
+    .cards-group-container .grouped-content-container {
+      grid-template-columns: repeat(3, 1fr);
+    }
+  }
+}
+
+
