--- aemedge/blocks/subnav/subnav.css
@@ -1,63 +1,0 @@
-.subnav-wrapper .subnav .subnav-group-container .grouped-content-container {
-  display: grid;
-  grid-template-columns: 1fr;
-  grid-auto-rows: auto;
-  gap: 40px;
-}
-
-.subnav-group-container .grouped-content-container .grouped-content {
-  border-left: 7px solid;
-  padding-left: 15px;
-  border-left-color: unset;
-  box-sizing: border-box;
-}
-
-.subnav-group-container .grouped-content-container .grouped-content p {
-  margin-top: 0;
-  font-family: Roboto, 'Roboto Fallback', sans-serif;
-  font-size: 14.4px;
-  color: #696969;
-}
-
-.subnav-group-container .grouped-content-container .grouped-content p:first-child {
-  font-weight: bold;
-  font-size: 22.4px;
-  color: var(--alt-link-hover-color);
-}
-
-.subnav-group-container .grouped-content-container .grouped-content p:first-child a {
-  color: var(--alt-link-color);
-  text-decoration: underline;
-  font-weight: bold;
-  font-size: 22.4px;
-}
-
-.subnav-group-container .grouped-content-container .grouped-content p a:hover {
-  text-decoration: none;
-  color: var(--alt-link-hover-color);
-}
-
-.grouped-content-container .grouped-content.blue {
-  border-left-color: #00607F;
-}
-
-.grouped-content-container .grouped-content.red {
-  border-left-color: #BB1F53;
-}
-
-.grouped-content-container .grouped-content.grey {
-    border-left-color: #B9C8D3;
-}
-
-@media (width >= 900px) {
-  .subnav-wrapper .subnav {
-    max-width: 950px;
-    margin: auto auto 40px 17%;
-
-    .subnav-group-container .grouped-content-container {
-      grid-template-columns: repeat(3, 1fr);
-    }
-  }
-}
-
-
