+++ aemedge/styles/styles.css
@@ -423,7 +423,7 @@ main .aside-details-wrapper {
     padding: 0 32px;
   }
 
-  .subnav-wrapper {
+  .cards-wrapper {
     margin-left: 12%;
   }
 }
